Bump CIRCT again, std dialect -> func dialect
