+++ src/circt-sys/build.rs
@@ -45,12 +45,13 @@ fn main() {
         "LLVMSupport",
         "MLIRAnalysis",
         "MLIRArithmetic",
+        "MLIRCAPIFunc",
         "MLIRCAPIIR",
         "MLIRCAPIControlFlow",
-        "MLIRCAPIStandard",
         "MLIRCallInterfaces",
         "MLIRControlFlow",
         "MLIRControlFlowInterfaces",
+        "MLIRFunc",
         "MLIRIR",
         "MLIRInferTypeOpInterface",
         "MLIRPDL",
@@ -60,7 +61,6 @@ fn main() {
         "MLIRPass",
         "MLIRRewrite",
         "MLIRSideEffectInterfaces",
-        "MLIRStandard",
         "MLIRSupport",
         "MLIRTransformUtils",
         "MLIRTransforms",
