+++ src/circt-sys/wrapper.h
@@ -12,13 +12,13 @@
 #include "mlir-c/Diagnostics.h"
 #include "mlir-c/Dialect/Async.h"
 #include "mlir-c/Dialect/ControlFlow.h"
+#include "mlir-c/Dialect/Func.h"
 #include "mlir-c/Dialect/GPU.h"
 #include "mlir-c/Dialect/LLVM.h"
 #include "mlir-c/Dialect/Linalg.h"
 #include "mlir-c/Dialect/SCF.h"
 #include "mlir-c/Dialect/Shape.h"
 #include "mlir-c/Dialect/SparseTensor.h"
-#include "mlir-c/Dialect/Standard.h"
 #include "mlir-c/Dialect/Tensor.h"
 #include "mlir-c/ExecutionEngine.h"
 #include "mlir-c/IR.h"
